+++ UnityMockTools/Assets/MSLib/Editor/Tools/UnitTestEditor/UnitTestEditorTestExecutionCallback.cs
@@ -2,7 +2,7 @@
 using UnityEditor.TestTools.TestRunner.Api;
 using UnityEngine;
 
-namespace MSLib
+namespace MSLib.Editor.Tools.UnitTestEditor
 {
     public sealed class UnitTestEditorTestExecutionCallback : ICallbacks
     {
