+++ UnityMockTools/Assets/MSLib/Editor/Tools/UnitTestEditor/UnitTestEditorWindow.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEditor;
 
-namespace MSLib
+namespace MSLib.Editor.Tools.UnitTestEditor
 {
     public sealed class UnitTestEditorWindow : EditorWindow
     {
